separate excel import logic from product agreement logic
